Moved some files around for cleanliness and deleted a file or two.
