+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/NKNOpMode_Basic.java
@@ -6,13 +6,12 @@ import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import com.qualcomm.robotcore.util.ElapsedTime;
 
 import org.nknsd.robotics.framework.NKNProgram;
-import org.nknsd.robotics.team.EACmonkeyProgram;
-import org.nknsd.robotics.team.ServoMonkeyProgram;
+import org.nknsd.robotics.team.BasicNKNProgram;
 
 import java.util.concurrent.TimeUnit;
 
-@TeleOp(name = "EAC Monkey")
-public class NKNOpMode_EACmonkey extends OpMode {
+@TeleOp(name = "Basic OpMode")
+public class NKNOpMode_Basic extends OpMode {
     private final ElapsedTime runtime = new ElapsedTime();
 
     //Time counters are in milliseconds
@@ -21,7 +20,7 @@ public class NKNOpMode_EACmonkey extends OpMode {
     // We use these two to delay telemetry outputs to ~200 milliseconds
 
     // Whatever program is attached here will be loaded with all its glorious components
-    private final NKNProgram program = new EACmonkeyProgram();
+    private final NKNProgram program = new BasicNKNProgram();
 
     // Code to run ONCE when the driver hits INIT
     @Override
