+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/autonomous/NKNAutonomous_Test.java
@@ -1,17 +1,17 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.autonomous;
 
 
+import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
-import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import com.qualcomm.robotcore.util.ElapsedTime;
 
 import org.nknsd.robotics.framework.NKNProgram;
-import org.nknsd.robotics.team.ServoMonkeyProgram;
+import org.nknsd.robotics.team.autonomous.PushAuto;
 
 import java.util.concurrent.TimeUnit;
 
-@TeleOp(name = "Servo Monkey")
-public class NKNOpMode_ServoMonkey extends OpMode {
+@Autonomous(name = "Autonomous Test")
+public class NKNAutonomous_Test extends OpMode {
     private final ElapsedTime runtime = new ElapsedTime();
 
     //Time counters are in milliseconds
@@ -20,19 +20,19 @@ public class NKNOpMode_ServoMonkey extends OpMode {
     // We use these two to delay telemetry outputs to ~200 milliseconds
 
     // Whatever program is attached here will be loaded with all its glorious components
-    private final NKNProgram program = new ServoMonkeyProgram();
+    private final NKNProgram program = new PushAuto();
 
     // Code to run ONCE when the driver hits INIT
     @Override
     public void init() {
         runtime.reset();
-        telemetry.addData("Status", "Initializing");
-        telemetry.update();
+        //telemetry.addData("Status", "Initializing");
+        //telemetry.update();
 
         program.init(telemetry, hardwareMap, gamepad1, gamepad2);
 
-        telemetry.addData("Status", "Initialized");
-        telemetry.update();
+        //telemetry.addData("Status", "Initialized");
+        //telemetry.update();
     }
 
     // Code to run REPEATEDLY after hitting INIT but before hitting PLAY
