+++ TeamCode/src/main/java/org/nknsd/robotics/team/monkeys/EACmonkeyProgram.java
@@ -1,24 +1,22 @@
-package org.nknsd.robotics.team;
+package org.nknsd.robotics.team.monkeys;
 
 import org.nknsd.robotics.framework.NKNComponent;
 import org.nknsd.robotics.framework.NKNProgram;
-import org.nknsd.robotics.team.components.PotentiometerHandler;
-import org.nknsd.robotics.team.components.testfiles.RotationMonkey;
 import org.nknsd.robotics.team.components.ExtensionHandler;
+import org.nknsd.robotics.team.components.IntakeSpinnerHandler;
+import org.nknsd.robotics.team.components.PotentiometerHandler;
 import org.nknsd.robotics.team.components.RotationHandler;
-
+import org.nknsd.robotics.team.components.testfiles.EACmonkey;
 
 import java.util.List;
 
-public class RotatorMonkeyProgram extends NKNProgram {
+public class EACmonkeyProgram extends NKNProgram {
     @Override
     public void createComponents(List<NKNComponent> components, List<NKNComponent> telemetryEnabled) {
-        // Create extension handler so that the rotator can be linked to it
         ExtensionHandler extensionHandler = new ExtensionHandler("motorArmExtend", true, 0.8);
         components.add(extensionHandler);
         telemetryEnabled.add(extensionHandler);
 
-        // Creating components we actually want
         PotentiometerHandler potentiometerHandler = new PotentiometerHandler("armPot");
         components.add(potentiometerHandler);
         telemetryEnabled.add(potentiometerHandler);
@@ -27,13 +25,16 @@ public class RotatorMonkeyProgram extends NKNProgram {
         components.add(rotationHandler);
         telemetryEnabled.add(rotationHandler);
 
-        RotationMonkey rotatorMonkey = new RotationMonkey();
-        rotatorMonkey.link(rotationHandler);
-        components.add(rotatorMonkey);
+        IntakeSpinnerHandler intakeSpinnerHandler = new IntakeSpinnerHandler("intakeServo");
+        components.add(intakeSpinnerHandler);
+        telemetryEnabled.add(intakeSpinnerHandler);
 
-        // Linking
+        EACmonkey eacMonkey = new EACmonkey();
+        components.add(eacMonkey);
+        telemetryEnabled.add(eacMonkey);
+
+        eacMonkey.link(extensionHandler, intakeSpinnerHandler, rotationHandler);
         rotationHandler.link(potentiometerHandler, extensionHandler);
         extensionHandler.link(rotationHandler);
-        telemetryEnabled.add(rotatorMonkey);
     }
 }
